Library

InfrastructureModels.jl Library

InfrastructureModels.update_data! — Method.

recursively applies new_data to data, overwriting information

source

InfrastructureModels._bold — Method.

Makes a string bold in the terminal

source

InfrastructureModels._float2string — Method.

converts a float value into a string of fixed precision

sprintf would do the job but this work around is needed because sprintf cannot take format strings during runtime

source

InfrastructureModels._grey — Method.

Makes a string grey in the terminal, does not seem to work well on Windows terminals more info can be found at https://en.wikipedia.org/wiki/ANSI_escape_code

source

InfrastructureModels._update_data! — Method.

recursive call of _update_data

source

InfrastructureModels._value2string — Method.

converts any value to a string, summarizes arrays and dicts

source

InfrastructureModels.add_line_delimiter — Method.

source

InfrastructureModels.check_type — Method.

Checks if the given value is of a given type, if not tries to make it that type

source

InfrastructureModels.extract_matlab_assignment — Method.

breaks up matlab strings of the form 'name = value;'

source

InfrastructureModels.parse_matlab_cells — Method.

source

InfrastructureModels.parse_matlab_data — Method.

source

InfrastructureModels.parse_matlab_matrix — Method.

source

InfrastructureModels.print_summary — Method.

prints the text summary for a data dictionary to STDOUT

source

InfrastructureModels.replicate — Method.

Transforms a single network into a multinetwork with several deepcopies of the original network

source

InfrastructureModels.split_line — Method.

source

InfrastructureModels.summary — Method.

prints the text summary for a data dictionary to IO

source

InfrastructureModels.type_array — Method.

Attempts to determine the type of an array of strings extracted from a matlab file

source

InfrastructureModels.type_value — Method.

Attempts to determine the type of a string extracted from a matlab file

source