Library

InfrastructureModels.jl Library

InfrastructureModels.update_data! — Method.

recursively applies new_data to data, overwriting information

source

InfrastructureModels._bold — Method.

Makes a string bold in the terminal

source

InfrastructureModels._float2string — Method.

converts a float value into a string of fixed precision

sprintf would do the job but this work around is needed because sprintf cannot take format strings during runtime

source

InfrastructureModels._grey — Method.

Makes a string grey in the terminal, does not seem to work well on Windows terminals more info can be found at https://en.wikipedia.org/wiki/ANSI_escape_code

source

InfrastructureModels._update_data! — Method.

recursive call of _update_data

source

InfrastructureModels._value2string — Method.

converts any value to a string, summarizes arrays and dicts

source

InfrastructureModels.add_line_delimiter — Method.

source

InfrastructureModels.arrays_to_dicts! — Method.

turns top level arrays into dicts

source

InfrastructureModels.check_type — Method.

Checks if the given value is of a given type, if not tries to make it that type

source

InfrastructureModels.compare_dict — Method.

tests if two dicts are equal, up to floating point precision

source

InfrastructureModels.compare_numbers — Method.

tests if two numbers are equal, up to floating point precision

source

InfrastructureModels.extract_matlab_assignment — Method.

breaks up matlab strings of the form 'name = value;'

source

InfrastructureModels.parse_matlab_cells — Method.

source

InfrastructureModels.parse_matlab_data — Method.

source

InfrastructureModels.parse_matlab_matrix — Method.

source

InfrastructureModels.print_summary — Method.

prints the text summary for a data dictionary to STDOUT

source

InfrastructureModels.relaxation_complex_product — Method.

constraint: c^2 + d^2 <= a*b

source

InfrastructureModels.relaxation_complex_product_on_off — Method.

c^2 + d^2 <= a*b*getupperbound(z)
c^2 + d^2 <= getupperbound(a)*b*getupperbound(z)
c^2 + d^2 <= a*getupperbound(b)*z

source

InfrastructureModels.relaxation_equality_on_off — Method.

x - getupperbound(x)*(1-z) <= y <= x - getlowerbound(x)*(1-z)

source

InfrastructureModels.relaxation_product — Method.

general relaxation of binlinear term (McCormick)

z >= getlowerbound(x)*y + getlowerbound(y)*x - getlowerbound(x)*getlowerbound(y)
z >= getupperbound(x)*y + getupperbound(y)*x - getupperbound(x)*getupperbound(y)
z <= getlowerbound(x)*y + getupperbound(y)*x - getlowerbound(x)*getupperbound(y)
z <= getupperbound(x)*y + getlowerbound(y)*x - getupperbound(x)*getlowerbound(y)

source

InfrastructureModels.relaxation_product_on_off — Method.

On/Off variant of binlinear term (McCormick) requires that all variables (x,y,z) go to zero with ind

source

InfrastructureModels.relaxation_sqr — Method.

general relaxation of a square term

x^2 <= y <= (getupperbound(x)+getlowerbound(x))*x - getupperbound(x)*getlowerbound(x)

source

InfrastructureModels.relaxation_trilinear — Method.

convex hull relaxation of trilinear term

w₁ = getlowerbound(x)*getlowerbound(y)*getlowerbound(z)
w₂ = getlowerbound(x)*getlowerbound(y)*getupperbound(z)
w₃ = getlowerbound(x)*getupperbound(y)*getlowerbound(z)
w₄ = getlowerbound(x)*getupperbound(y)*getupperbound(z)
w₅ = getupperbound(x)*getlowerbound(y)*getlowerbound(z)
w₆ = getupperbound(x)*getlowerbound(y)*getupperbound(z)
w₇ = getupperbound(x)*getupperbound(y)*getlowerbound(z)
w₈ = getupperbound(x)*getupperbound(y)*getupperbound(z)
w = λ₁*w₁ + λ₂*w₂ + λ₃*w₃ + λ₄*w₄ + λ₅*w₅ + λ₆*w₆ + λ₇*w₇ + λ₈*w₈
x = (λ₁ + λ₂ + λ₃ + λ₄)*getlowerbound(x) + (λ₅ + λ₆ + λ₇ + λ₈)*getupperbound(x)
y = (λ₁ + λ₂ + λ₅ + λ₆)*getlowerbound(x) + (λ₃ + λ₄ + λ₇ + λ₈)*getupperbound(x)
z = (λ₁ + λ₃ + λ₅ + λ₇)*getlowerbound(x) + (λ₂ + λ₄ + λ₆ + λ₈)*getupperbound(x)
λ₁ + λ₂ + λ₃ + λ₄ + λ₅ + λ₆ + λ₇ + λ₈ = 1

source

InfrastructureModels.replicate — Method.

Transforms a single network into a multinetwork with several deepcopies of the original network

source

InfrastructureModels.split_line — Method.

source

InfrastructureModels.summary — Method.

prints the text summary for a data dictionary to IO

source

InfrastructureModels.type_array — Method.

Attempts to determine the type of an array of strings extracted from a matlab file

source

InfrastructureModels.type_value — Method.

Attempts to determine the type of a string extracted from a matlab file

source