Library

InfrastructureModels.jl Library

InfrastructureModels.update_data! — Method.

recursively applies new_data to data, overwriting information

source

InfrastructureModels._bold — Method.

Makes a string bold in the terminal

source

InfrastructureModels._float2string — Method.

converts a float value into a string of fixed precision

sprintf would do the job but this work around is needed because sprintf cannot take format strings during runtime

source

InfrastructureModels._grey — Method.

Makes a string grey in the terminal, does not seem to work well on Windows terminals more info can be found at https://en.wikipedia.org/wiki/ANSIescapecode

source

InfrastructureModels._iscomponentdict — Method.

Attempts to determine if the given data is a component dictionary

source

InfrastructureModels._update_data! — Method.

recursive call of updatedata

source

InfrastructureModels._value2string — Method.

converts any value to a string, summarizes arrays and dicts

source

InfrastructureModels.add_line_delimiter — Method.

source

InfrastructureModels.arrays_to_dicts! — Method.

turns top level arrays into dicts

source

InfrastructureModels.check_type — Method.

Checks if the given value is of a given type, if not tries to make it that type

source

InfrastructureModels.compare_dict — Method.

tests if two dicts are equal, up to floating point precision

source

InfrastructureModels.compare_numbers — Method.

tests if two numbers are equal, up to floating point precision

source

InfrastructureModels.component_table — Method.

builds a table of component data

source

InfrastructureModels.extract_matlab_assignment — Method.

breaks up matlab strings of the form 'name = value;'

source

InfrastructureModels.ismultinetwork — Method.

checks if a given network data is a multinetwork

source

InfrastructureModels.parse_matlab_cells — Method.

source

InfrastructureModels.parse_matlab_data — Method.

source

InfrastructureModels.parse_matlab_matrix — Method.

source

InfrastructureModels.print_summary — Method.

prints the text summary for a data dictionary to stdout

source

InfrastructureModels.relaxation_complex_product — Method.

constraint: c^2 + d^2 <= a*b

source

InfrastructureModels.relaxation_complex_product_conic — Method.

a conic encoding of constraint: c^2 + d^2 <= a*b

source

InfrastructureModels.relaxation_complex_product_on_off — Method.

c^2 + d^2 <= a*b*JuMP.upper_bound(z)
c^2 + d^2 <= JuMP.upper_bound(a)*b*JuMP.upper_bound(z)
c^2 + d^2 <= a*JuMP.upper_bound(b)*z

source

InfrastructureModels.relaxation_equality_on_off — Method.

x - JuMP.upper_bound(x)*(1-z) <= y <= x - JuMP.lower_bound(x)*(1-z)

source

InfrastructureModels.relaxation_product — Method.

general relaxation of binlinear term (McCormick)

z >= JuMP.lower_bound(x)*y + JuMP.lower_bound(y)*x - JuMP.lower_bound(x)*JuMP.lower_bound(y)
z >= JuMP.upper_bound(x)*y + JuMP.upper_bound(y)*x - JuMP.upper_bound(x)*JuMP.upper_bound(y)
z <= JuMP.lower_bound(x)*y + JuMP.upper_bound(y)*x - JuMP.lower_bound(x)*JuMP.upper_bound(y)
z <= JuMP.upper_bound(x)*y + JuMP.lower_bound(y)*x - JuMP.upper_bound(x)*JuMP.lower_bound(y)

source

InfrastructureModels.relaxation_product_on_off — Method.

On/Off variant of binlinear term (McCormick) requires that all variables (x,y,z) go to zero with ind

source

InfrastructureModels.relaxation_sqr — Method.

general relaxation of a square term

x^2 <= y <= (JuMP.upper_bound(x)+JuMP.lower_bound(x))*x - JuMP.upper_bound(x)*JuMP.lower_bound(x)

source

InfrastructureModels.relaxation_trilinear — Method.

convex hull relaxation of trilinear term

w₁ = JuMP.lower_bound(x)*JuMP.lower_bound(y)*JuMP.lower_bound(z)
w₂ = JuMP.lower_bound(x)*JuMP.lower_bound(y)*JuMP.upper_bound(z)
w₃ = JuMP.lower_bound(x)*JuMP.upper_bound(y)*JuMP.lower_bound(z)
w₄ = JuMP.lower_bound(x)*JuMP.upper_bound(y)*JuMP.upper_bound(z)
w₅ = JuMP.upper_bound(x)*JuMP.lower_bound(y)*JuMP.lower_bound(z)
w₆ = JuMP.upper_bound(x)*JuMP.lower_bound(y)*JuMP.upper_bound(z)
w₇ = JuMP.upper_bound(x)*JuMP.upper_bound(y)*JuMP.lower_bound(z)
w₈ = JuMP.upper_bound(x)*JuMP.upper_bound(y)*JuMP.upper_bound(z)
w = λ₁*w₁ + λ₂*w₂ + λ₃*w₃ + λ₄*w₄ + λ₅*w₅ + λ₆*w₆ + λ₇*w₇ + λ₈*w₈
x = (λ₁ + λ₂ + λ₃ + λ₄)*JuMP.lower_bound(x) + (λ₅ + λ₆ + λ₇ + λ₈)*JuMP.upper_bound(x)
y = (λ₁ + λ₂ + λ₅ + λ₆)*JuMP.lower_bound(y) + (λ₃ + λ₄ + λ₇ + λ₈)*JuMP.upper_bound(y)
z = (λ₁ + λ₃ + λ₅ + λ₇)*JuMP.lower_bound(z) + (λ₂ + λ₄ + λ₆ + λ₈)*JuMP.upper_bound(z)
λ₁ + λ₂ + λ₃ + λ₄ + λ₅ + λ₆ + λ₇ + λ₈ = 1

source

InfrastructureModels.replicate — Method.

Transforms a single network into a multinetwork with several deepcopies of the original network

source

InfrastructureModels.row_to_dict — Method.

takes a row from a matrix and assigns the values names

source

InfrastructureModels.row_to_typed_dict — Method.

takes a row from a matrix and assigns the values names and types

source

InfrastructureModels.split_line — Method.

source

InfrastructureModels.summary — Method.

prints the text summary for a data dictionary to IO

source

InfrastructureModels.type_array — Method.

Attempts to determine the type of an array of strings extracted from a matlab file

source

InfrastructureModels.type_value — Method.

Attempts to determine the type of a string extracted from a matlab file

source

InfrastructureModels.variable_domain — Method.

Computes the valid domain of a given JuMP variable taking into account bounds and the varaible's implicit bounds (e.g. binary).

source